+++ webextension/js/init-templates.js
@@ -2,6 +2,7 @@ const templatesSource = window.templatesSource = new Map();
 templatesSource.set('backgroundTheme', '/templates/backgroundTheme');
 templatesSource.set('panelCheckedDataItem', '/templates/panel/checkedDataItem');
 templatesSource.set('tabMover', '/templates/panel/tabMover');
+templatesSource.set('tabMoverNative', '/templates/panel/tabMoverNative');
 templatesSource.set('tabPageServerIp', '/templates/panel/tabPageServerIp');